typesの配下にあったchord.tsをdata配下に移動し、ファイルを分割
