+++ src/data/chord/interfaces.ts
@@ -1,0 +1,8 @@
+/**
+ * コード情報
+ */
+export interface ChordParts {
+    keyNote: string; // キー音
+    modifier: string; // コード修飾
+    bassNote: string; // ベース音（オプション）
+}
