+++ src/types/index.ts
@@ -1,6 +1,5 @@
 export * from "./commons/APIState";
 export * from "./commons/SelectOption";
-export * from "./commons/chord";
 
 export * from "./enums/SearchType";
 export * from "./enums/CapoValue";
